Web console: refactor home view, add tests (#8247)
* refactor home view

* updated mode button placement
