+++ web-console/src/views/home-view/home-view-card/home-view-card.scss
@@ -16,15 +16,8 @@
  * limitations under the License.
  */
 
-@import '../../variables';
-
-.home-view {
-  display: grid;
-  grid-gap: $standard-padding;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-
-  & > a {
-    text-decoration: inherit;
-    color: inherit;
+.home-view-card {
+  .bp3-card {
+    height: 170px;
   }
 }
